backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
PyPDF2
langchain
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
Pillow
celery
redis
